requirements.txt
Django==1.6.2
Pillow==2.4.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-bootstrap-toolkit==2.15.0
django-bootstrap3==4.2.0
django-tastypie==0.11.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
python-mimeparse==0.1.4
static==1.0.2
wsgiref==0.1.2
